requirements.txt
pandas
yfinance
python-dotenv